feat(scrcpy): extract video codec parsers to own package
use namespace exports for better name separation
move `toCodecString` methods to codec parsers
