+++ libraries/adb-scrcpy/package.json
@@ -36,6 +36,7 @@
         "@yume-chan/adb": "workspace:^",
         "@yume-chan/async": "^4.1.3",
         "@yume-chan/event": "workspace:^",
+        "@yume-chan/media-codec": "workspace:^",
         "@yume-chan/scrcpy": "workspace:^",
         "@yume-chan/stream-extra": "workspace:^",
         "@yume-chan/struct": "workspace:^"
