+++ libraries/adb-server-node-tcp/tsconfig.build.json
@@ -9,12 +9,4 @@
             "node"
         ]
     },
-    "references": [
-        {
-            "path": "../adb/tsconfig.build.json"
-        },
-        {
-            "path": "../stream-extra/tsconfig.build.json"
-        }
-    ]
 }
