+++ libraries/android-bin/tsconfig.json
@@ -5,6 +5,15 @@
         },
         {
             "path": "./tsconfig.test.json"
+        },
+        {
+            "path": "../adb/tsconfig.build.json"
+        },
+        {
+            "path": "../stream-extra/tsconfig.build.json"
+        },
+        {
+            "path": "../struct/tsconfig.build.json"
         }
     ]
 }
