+++ libraries/media-codec/.npmignore
@@ -1,0 +1,16 @@
+.rush
+
+# Test
+coverage
+**/*.spec.ts
+**/*.spec.js
+**/*.spec.js.map
+**/__helpers__
+jest.config.js
+
+.eslintrc.cjs
+tsconfig.json
+tsconfig.test.json
+
+# Logs
+*.log
