+++ libraries/media-codec/README.md
@@ -1,0 +1,3 @@
+# @yume-chan/media-codec
+
+H.264, H.265 and AV1 configuration packet parser
