+++ libraries/media-codec/tsconfig.build.json
@@ -1,0 +1,3 @@
+{
+    "extends": "./node_modules/@yume-chan/tsconfig/tsconfig.base.json"
+}
