+++ libraries/media-codec/tsconfig.test.json
@@ -1,0 +1,9 @@
+{
+    "extends": "./tsconfig.build.json",
+    "compilerOptions": {
+        "types": [
+            "node"
+        ],
+    },
+    "exclude": []
+}
