+++ libraries/scrcpy-decoder-tinyh264/package.json
@@ -35,6 +35,7 @@
     "dependencies": {
         "@yume-chan/async": "^4.1.3",
         "@yume-chan/event": "workspace:^",
+        "@yume-chan/media-codec": "workspace:^",
         "@yume-chan/scrcpy": "workspace:^",
         "@yume-chan/stream-extra": "workspace:^",
         "tinyh264": "^0.0.7",
