+++ libraries/scrcpy-decoder-tinyh264/tsconfig.json
@@ -5,6 +5,15 @@
         },
         {
             "path": "./tsconfig.test.json"
-        }
+        },
+        {
+            "path": "../media-codec/tsconfig.build.json"
+        },
+        {
+            "path": "../scrcpy/tsconfig.build.json"
+        },
+        {
+            "path": "../stream-extra/tsconfig.build.json"
+        },
     ]
 }
