+++ libraries/scrcpy-decoder-webcodecs/src/video/codec/h265.ts
@@ -1,8 +1,6 @@
-import { getUint32LittleEndian } from "@yume-chan/no-data-view";
-import { h265ParseConfiguration } from "@yume-chan/scrcpy";
+import { H265 } from "@yume-chan/media-codec";
 
 import { H26xDecoder } from "./h26x.js";
-import { hexDigits } from "./utils.js";
 
 export class H265Decoder extends H26xDecoder {
     #decoder: VideoDecoder;
@@ -18,32 +16,18 @@ export class H265Decoder extends H26xDecoder {
     }
 
     override configure(data: Uint8Array): void {
-        const {
-            generalProfileSpace,
-            generalProfileIndex,
-            generalProfileCompatibilitySet,
-            generalTierFlag,
-            generalLevelIndex,
-            generalConstraintSet,
-            croppedWidth,
-            croppedHeight,
-        } = h265ParseConfiguration(data);
+        const configuration = H265.parseConfiguration(data);
 
-        this.#updateSize(croppedWidth, croppedHeight);
+        this.#updateSize(
+            configuration.croppedWidth,
+            configuration.croppedHeight,
+        );
 
-        const codec = [
-            "hev1",
-            ["", "A", "B", "C"][generalProfileSpace]! +
-                generalProfileIndex.toString(),
-            hexDigits(getUint32LittleEndian(generalProfileCompatibilitySet, 0)),
-            (generalTierFlag ? "H" : "L") + generalLevelIndex.toString(),
-            ...Array.from(generalConstraintSet, hexDigits),
-        ].join(".");
         this.#decoder.configure({
-            codec,
+            codec: H265.toCodecString(configuration),
             // Microsoft Edge requires explicit size to work
-            codedWidth: croppedWidth,
-            codedHeight: croppedHeight,
+            codedWidth: configuration.croppedWidth,
+            codedHeight: configuration.croppedHeight,
             optimizeForLatency: true,
         });
     }
