+++ libraries/scrcpy-decoder-webcodecs/tsconfig.build.json
@@ -7,18 +7,4 @@
         ],
         "types": []
     },
-    "references": [
-        {
-            "path": "../no-data-view/tsconfig.build.json"
-        },
-        {
-            "path": "../scrcpy-decoder-tinyh264/tsconfig.build.json"
-        },
-        {
-            "path": "../scrcpy/tsconfig.build.json"
-        },
-        {
-            "path": "../stream-extra/tsconfig.build.json"
-        }
-    ]
 }
