+++ libraries/scrcpy-decoder-webcodecs/tsconfig.json
@@ -5,6 +5,21 @@
         },
         {
             "path": "./tsconfig.test.json"
+        },
+        {
+            "path": "../media-codec/tsconfig.build.json"
+        },
+        {
+            "path": "../no-data-view/tsconfig.build.json"
+        },
+        {
+            "path": "../scrcpy-decoder-tinyh264/tsconfig.build.json"
+        },
+        {
+            "path": "../scrcpy/tsconfig.build.json"
+        },
+        {
+            "path": "../stream-extra/tsconfig.build.json"
         }
     ]
 }
