--- libraries/scrcpy/src/codec/index.ts
@@ -1,4 +1,0 @@
-export * from "./av1.js";
-export * from "./h264.js";
-export * from "./h265.js";
-export * from "./nalu.js";
