+++ libraries/scrcpy/src/video/android.ts
@@ -1,0 +1,72 @@
+// From https://developer.android.com/reference/android/media/MediaCodecInfo.CodecProfileLevel
+export const AndroidAvcProfile = {
+    Baseline: 1 << 0,
+    Main: 1 << 1,
+    Extended: 1 << 2,
+    High: 1 << 3,
+    High10: 1 << 4,
+    High422: 1 << 5,
+    High444: 1 << 6,
+    ConstrainedBaseline: 1 << 16,
+    ConstrainedHigh: 1 << 19,
+};
+
+export const AndroidAvcLevel = {
+    Level1: 1 << 0,
+    Level1b: 1 << 1,
+    Level11: 1 << 2,
+    Level12: 1 << 3,
+    Level13: 1 << 4,
+    Level2: 1 << 5,
+    Level21: 1 << 6,
+    Level22: 1 << 7,
+    Level3: 1 << 8,
+    Level31: 1 << 9,
+    Level32: 1 << 10,
+    Level4: 1 << 11,
+    Level41: 1 << 12,
+    Level42: 1 << 13,
+    Level5: 1 << 14,
+    Level51: 1 << 15,
+    Level52: 1 << 16,
+    Level6: 1 << 17,
+    Level61: 1 << 18,
+    Level62: 1 << 19,
+};
+
+export const AndroidHevcProfile = {
+    Main: 1 << 0,
+    Main10: 1 << 1,
+    MainStill: 1 << 2,
+    Main10Hdr10: 1 << 12,
+    Main10Hdr10Plus: 1 << 13,
+};
+
+export const AndroidHevcLevel = {
+    MainTierLevel1: 1 << 0,
+    HighTierLevel1: 1 << 1,
+    MainTierLevel2: 1 << 2,
+    HighTierLevel2: 1 << 3,
+    MainTierLevel21: 1 << 4,
+    HighTierLevel21: 1 << 5,
+    MainTierLevel3: 1 << 6,
+    HighTierLevel3: 1 << 7,
+    MainTierLevel31: 1 << 8,
+    HighTierLevel31: 1 << 9,
+    MainTierLevel4: 1 << 10,
+    HighTierLevel4: 1 << 11,
+    MainTierLevel41: 1 << 12,
+    HighTierLevel41: 1 << 13,
+    MainTierLevel5: 1 << 14,
+    HighTierLevel5: 1 << 15,
+    MainTierLevel51: 1 << 16,
+    HighTierLevel51: 1 << 17,
+    MainTierLevel52: 1 << 18,
+    HighTierLevel52: 1 << 19,
+    MainTierLevel6: 1 << 20,
+    HighTierLevel6: 1 << 21,
+    MainTierLevel61: 1 << 22,
+    HighTierLevel61: 1 << 23,
+    MainTierLevel62: 1 << 24,
+    HighTierLevel62: 1 << 25,
+};
