+++ libraries/scrcpy/src/video/size.ts
