+++ libraries/scrcpy/tsconfig.build.json
@@ -1,11 +1,3 @@
 {
     "extends": "./node_modules/@yume-chan/tsconfig/tsconfig.base.json",
-    "references": [
-        {
-            "path": "../stream-extra/tsconfig.build.json"
-        },
-        {
-            "path": "../struct/tsconfig.build.json"
-        }
-    ]
 }
