+++ libraries/scrcpy/tsconfig.json
@@ -5,6 +5,12 @@
         },
         {
             "path": "./tsconfig.test.json"
+        },
+        {
+            "path": "../stream-extra/tsconfig.build.json"
+        },
+        {
+            "path": "../struct/tsconfig.build.json"
         }
     ]
 }
